a_star as an isolated module
